pyproject.toml
 [build-system]
 requires = ["setuptools>=68", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "novelagent"
 version = "0.1.0"
 description = "A local CLI novel editing-department agent with continuous learning."
 readme = "README.md"
 requires-python = ">=3.10"
 authors = [{ name = "50pai" }]
 dependencies = [
   "pydantic>=2.6",
   "typer>=0.12",
   "rich>=13.7",
   "pyyaml>=6.0",
   "numpy>=1.26",
   "tiktoken>=0.7.0",
   "httpx>=0.27",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
 ]

[project.optional-dependencies]
vectorstore = [
  "chromadb>=0.5.5",
]
 
 [project.scripts]
 novelagent = "novelagent.cli:app"
 
 [tool.setuptools]
 package-dir = {"" = "src"}
 
 [tool.setuptools.packages.find]
 where = ["src"]
 
 [tool.ruff]
 line-length = 100
 target-version = "py310"
 
